Add and apply aspect ratio classes. Target #card-container instead of #resizable
The reason it worked at the very beginning was because we were resizing the width of card-container, and filled it with a credit card image that had a constant aspect ratio. We then replaced the image with a credit card div, which had an assigned width/height. When we tried to target the #resizable credit card div instead, it did not scale with a constant aspect ratio because we were resizing only its width. Now, we're going back and resizing the container's width, ensuring that the aspect ratio is constant with the new .credit-card-aspect-ratio class, and filling up the space by using .aspect-content, which absolutely positions the contents of the card-container (the credit card itself).

TL;DR Image resizing != div resizing